refactor: api, html 변경등 파일 분리 및 폴더 구조 변경
