+++ ssr/server/utils/round.js
@@ -1,0 +1,7 @@
+const round = (value, decimals = 0) => {
+  const factor = 10 ** decimals;
+
+  return Math.round(value * factor) / factor;
+};
+
+export default round;
